Refactor beforeEach and afterEach helpers into their own file
